This is the second commit
It appears at this point in time that the bubble sorter works.  There are some fringe use cases that you'll want to pay attention to, like it won't work at all if the array is not of an appropriate size, and conceptually there is room to program those,  but since we define the array at the beginning,  the more sensible solution is to make sure you define your array as being more than 2 in length.